requirements.txt
aiohttp
python-telegram-bot
pandas
beautifulsoup4
requests
python-dotenv